requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bitsandbytes==0.45.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
cut-cross-entropy==25.1.1
datasets==3.2.0
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring-parser==0.16
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.14.0
greenlet==3.1.1
h11==0.14.0
hf-transfer==0.1.9
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib-metadata==8.6.1
jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama-cpp-python==0.3.7
loguru==0.7.3
mako==1.3.9
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.61.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==11.1.0
pip==25.0
propcache==0.2.1
protobuf==3.20.3
psutil==7.0.0
pyarrow==19.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyyaml==6.0.2
ray==2.42.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
sentencepiece==0.2.0
setuptools==75.8.0
shtab==1.7.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.38
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
trl==0.14.0
typeguard==4.4.1
typing-extensions==4.12.2
tyro==0.9.14
tzdata==2025.1
unsloth==2025.2.9
unsloth-zoo==2025.2.4
urllib3==2.3.0
wheel==0.45.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
